yeah

+++ src/frame.cpp
@@ -6,7 +6,7 @@ funkin::Frame::Frame(raylib::Rectangle rect)
 {
 }
 
-funkin::Frame::Frame(int x, int y, int width, int height)
+funkin::Frame::Frame(float x, float y, float width, float height)
 {
     this->x = x;
     this->y = y;
